WIP on api route not requiring node
