+++ packages/keystatic/package.json
@@ -284,6 +284,10 @@
       "node": "./src/sha1/node.ts",
       "default": "./src/sha1/webcrypto.ts"
     },
+    "#webcrypto": {
+      "node": "./src/api/webcrypto/node.ts",
+      "default": "./src/api/webcrypto/webcrypto.ts"
+    },
     "#field-ui/*": {
       "react-server": "./src/form/fields/empty-field-ui.tsx",
       "default": "./src/form/fields/*/ui.tsx"
