+++ packages/keystatic/src/api/webcrypto/default.ts
@@ -1,0 +1,1 @@
+export const webcrypto = crypto;
