+++ packages/keystatic/src/api/webcrypto/node.ts
@@ -1,0 +1,1 @@
+export { webcrypto } from 'node:crypto';
