+++ packages/keystatic/src/sha1/webcrypto.ts
@@ -1,8 +1,6 @@
+import { bytesToHex } from '../hex';
+
 export async function sha1(content: Uint8Array) {
   const hashBuffer = await crypto.subtle.digest('SHA-1', content);
-  let str = '';
-  for (const byte of new Uint8Array(hashBuffer)) {
-    str += byte.toString(16).padStart(2, '0');
-  }
-  return str;
+  return bytesToHex(new Uint8Array(hashBuffer));
 }
